// apps/consumer-experience/src/components/add-bank/shared-types.ts
import { AccountType } from '@zinnia/api-types/types/sor';

export enum FormMode {
  EDIT = 'edit',
  ADD = 'add',
}

export enum FormSteps {
  ADD = 'ADD',
  LOADING = 'LOADING',
  ERROR = 'ERROR',
  SUCCESS = 'SUCCESS',
}

export interface BankFormFields {
  accountType?: AccountType;
  branchName?: string;
  routingNumber?: string;
  accountNumber?: string;
  autopayEnabled?: boolean;
}